+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Allow custom aria attributs in Rating component
+- SONAR-13553 Do not allow multiple clicks on 'Load more' link in ListFooter
+- ListFooter "Load more" and "Reload" links are now buttons
 
 ## 1.0.10